refactor: LottoGenerator를 구현하는 RandomGenerator와 FixedGenerator 클래스 생성
